requirements.txt
torch
torchvision

# For Testing
numpy
pillow
git+https://github.com/FabianHertwig/torch_videovision